--- app/views/layouts/application_single_colum.html.haml
@@ -1,10 +1,0 @@
-!!!
-%html
-  %head
-    =render :partial => 'layouts/layout_header'
-  %body
-    =render :partial => 'layouts/layout_nav_bar'
-    .container.pull-down.with-padding
-      =render :partial => 'shared/flash_messages'
-      = yield
-    =render :partial => 'layouts/layout_footer'
