+++ app/views/v2014/articles/edit.html.haml
@@ -4,4 +4,4 @@
 
 = link_to 'Show', @article
 \|
-= link_to 'Back', articles_path
+= link_to 'Back', v2014_articles_path
