+++ app/views/v2014/articles/index.json.jbuilder
