+++ app/views/v2014/articles/show.html.haml
@@ -10,4 +10,4 @@
       = @article.added_by
 
 
-= link_to 'Back', articles_path
+= link_to 'Back', v2014_articles_path
