+++ app/views/v2014/articles/show.json.jbuilder
