+++ app/views/v2014/post/faq.html.haml
@@ -6,7 +6,7 @@
       %br
       %br
       Please find a list of questions that we are frequently asked. Feel free to
-      %a{href: "/contact/new", title: "Contact Us"} contact us
+      %a{href: "/v2014/contact_us", title: "Contact Us"} contact us
       in case you dont find the answers satisfactory or have a question which was not addressed below.
     %h3.green Regarding Course Content
     #coursegroup.panel-group
