+++ app/views/v2014/post/gallery.html.haml
