+++ app/views/v2014/post/index.html.haml
@@ -6,7 +6,7 @@
         %span.text-muted Dayananda Sagar College of Engineering
       %p.lead 
         FSMK is glad to announce that Information Science Department of Dayananda Sagar College of Engineering has agreed to be the co-organizer and host the summer camp in their Kumaraswamy Layout Campus. Please check
-        %a{:href => venue_path} the venue page
+        %a{:href => v2014_venue_path} the venue page
         for more information.
     .col-md-5
       = image_tag "college-logo.png", alt: "Dayananda Sagar College Of Engineering", :class => "featurette-image img-responsive"
@@ -28,15 +28,15 @@
   / Three columns of text below the carousel
   .row
     .col-sm-4
-      %a.btn{:href => speakers_path, :role => "button"}
+      %a.btn{:href => v2014_speakers_path, :role => "button"}
         %h2.green Speakers
         =image_tag "speakers.png", alt: "Speakers", :class => "img-circle"  
     .col-sm-4
-      %a.btn{:href => technologies_path, :role => "button"}
+      %a.btn{:href => v2014_technologies_path, :role => "button"}
         %h2.green Technologies
         =image_tag "technologies.png", alt: "Technologies", :class => "img-circle"
     .col-sm-4
-      %a.btn{:href => sponsors_path, :role => "button"}
+      %a.btn{:href => v2014_sponsors_path, :role => "button"}
         %h2.green Sponsors
         =image_tag "sponsorship.png", alt: "Sponsorship", :class => "img-circle"
