+++ app/views/v2014/post/speakers.html.haml
