+++ app/views/v2014/post/sponsors.html.haml
