+++ app/views/v2014/post/venue.html.haml
@@ -30,9 +30,9 @@
       = "Get in touch"
     %p.lead
       Please do 
-      %a{:href => contact_us_path} Contact Us
+      %a{:href => v2014_contact_us_path} Contact Us
       in case you have any doubts regarding the venue. Also check the 
-      %a{:href => faq_path} FAQs Page
+      %a{:href => v2014_faq_path} FAQs Page
       for frequently asked questions regarding the hostel facilities.
   .col-md-5
     #map
