+++ app/views/v2014/shared/_remaining_day_for_camp.html.haml
