+++ app/views/v2014/users/new.html.haml
@@ -23,7 +23,7 @@
         %br
         .jumbotron.text-center
           %h4.date 2014 July 19 to 27
-    =render :partial => 'shared/remaining_day_for_camp'
+    =render :partial => 'v2014/shared/remaining_day_for_camp'
 - if @user.errors.any?
   :javascript
     $(document).on('ready page:load', function(){
