+++ app/views/v2014/volunteers/_form.html.haml
@@ -1,4 +1,4 @@
-= render partial: 'shared/error_messages', locals: { object: @volunteer }
+= render partial: 'v2014/shared/error_messages', locals: { object: @volunteer }
 = form_for @volunteer, :html => {role: "form"}  do |f|
   .form-group
     =f.label :name
